Commit no1
SplashFragment
onboarding - FirstScreen
             SecondScreen
             ThirdScreen
profileimage.xml

+++ app/src/main/java/com/example/oncash/Fragment/redeem.kt
@@ -90,14 +90,14 @@ class redeem : Fragment() {
             binding.withdrawButton.setOnClickListener {
                 val requestAmount = binding.withdrawRequestedAmount.text.toString()
                 if (requestAmount.isNotEmpty()) {
-                    if (walletBalance.toInt() >= requestAmount.toInt()) {
+                    if (walletBalance >= requestAmount.toInt()) {
                         if (requestAmount.toInt() > 20) {
 
                             viewModel.withdrawRequest(
                                 userNumber,
                                 requestAmount.toInt(),
                                 walletBalance,
-                                userRecordId  ,
+                                userRecordId,
                             )
                             viewModel.getWithdrawalRequest().observe(viewLifecycleOwner) { status ->
                                 if (status.response.contains("200")) {
